requirements.txt
Flask
gunicorn
Pillow
tqdm